requirements.txt
tensorflow==2.16.1
opencv-python==4.10.0.84
numpy==1.26.4
scikit-learn==1.4.2
matplotlib==3.8.4
pillow==10.3.0
pandas==2.2.2
tqdm==4.66.4
paho-mqtt==1.6.1
sounddevice==0.4.6
scipy==1.13.1
kaggle==1.6.17